detect_document/requirements.txt
boto3
jeffy
google-cloud-vision
